+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.2.9]
+* Add more detailed error codes for JIT (#1639)
+* Add support for nested auth protocol (#1631)
+
 ## [1.2.8]
 * Fix a crash when identity is nil during getting device registration information on iOS (#1184)